qgis-test-run-algorithm.R: minimize top-level code *
Leaving the input object in the top-level code however; it is used
to generate the test titles appropriately. Its state will be
changed if and when the same tests are re-run by
qgis-test-run-algorithm2.R. The latter is run to check with the
non-default input method, that is typically used on pre-3.23 QGIS
versions.