GHA R-CMD-check: postpone R dependencies *
It seems that system dependencies of the R package deps conflict with those of
ubuntugis-unstable PPA when QGIS is installed from the latter, at a
later stage.

This may be the cause of QGIS installation error in the ubuntugis job.

So first activating the PPA, then installing QGIS and finally the R pkg deps
seems a safer practice.