Drop rgdal and fix some raster::crs() tests *
Contrary to previous GDAL>=3 behaviour, raster::crs() now yields a very
narrow object with proj.4 string; it can't even retrieve the WKT anymore.

So adapted the test to make it succeed; to be seen in future whether new
changes are needed.

Anyhow, raster::crs() seems not to depend on rgdal anymore, so it's safe to
drop rgdal, which is needed anyway since it will retire this year.