test-compat-raster.R: use raster::wkt() *
When doing 94de816, I didn't manage to get the WKT. It seems I was using an older,
separate installation of raster at the time.

With current raster, this poses no problem. So this reverts the approach of manually
extracting the projargs slot and replaces raster::crs() by raster::wkt() as it should
be.