GHA R-CMD-check: install SAGA NextGen plugin in Linux *
It is needed to build the 'getting started' vignette.

I did not take the time to investigate how to download and unzip files in Win and macOS and do the same there.
So the vignette remains largely untested in Win & macOS (most of its code is in a child document, triggered by
the availability of the plugin)