Move inst/algorithms to misc/algorithms *
This directory is included in the source code, but is merely a manual
snapshot of the developer's local state of QGIS, for development
purposes.

These relatively large files should not be included in the installed package.
This commit takes the first step to do that, i.e. exclude from inst/.
Next step is to exclude from installing.

To be revisited for the sake of the source code itself. The files were
last updated Oct 8 2020.