Further handling of spDataLarge not being on CRAN *
Besides the Additional_repositories field in DESCRIPTION (0da526f),
care is also needed for two other points:

- since the Additional_repositories field (currently) is not used by pak to install
spDataLarge from, nor by remotes::install_github(), its installation in GitHub Actions is
to be taken care of using the extra-repositories field of the r-setup action. This should
make the vignette compile without error.

- for local installation in absence of spDataLarge and the extra repository, with building
vignettes, the spDataLarge availability is added to the condition for building the main
vignette.