roxygen docs: implement '@examplesIf has_qgis()' (thanks @paleolimbot) *
Note that this is not only to please the user visually.

More importantly, the examples run in pkgdown only show the last result in
case of multiple lines included in `if () {...}`.

So the only alternative is to use if (has_qgis()) on every line, as has been
done in qgis_has() documentation.