GHA R-CMD-check: drop R-devel job on macOS *
R packages on CRAN are not built for R-devel on macOS, contrary to Windows.
This makes the building process last very long.
Moreover, currently it poses a problem with building terra.

Most importantly, CRAN submissions appear not tested against that
combination (https://cran.r-project.org/web/checks/check_flavors.html).

For these reasons, the job is dropped and it is assumed that the terra package
will get a binary build on CRAN for macOS, at the next R release.