Dynamically register sf/stars S3 methods in NAMESPACE, not with vctrs::s3_register() *
Relates to CRAN review: https://github.com/r-spatial/qgisprocess/issues/167#issuecomment-1680128232.

Before, vctrs::s3_register() was used to dynamically register S3 methods for
generics in suggested packages, in this case for sf::st_as_sf() and stars::st_as_stars().

However we also want to have these methods documented, with examples, and appearing
in the package index. This was already the case, but CRAN requires examples to be
present only for exported methods (NAMESPACE).

Delayed S3 method registration is supported since R>=3.6.0, and can be implemented with
the roxygen2 tag @exportS3Method.

The downside of this non-vctrs implementation is the minimum required R version.