qgis_reconfigure(): allow for incomplete configuration without generating error *
Avoiding to throw an error during configuration also avoids triggering qgis_unconfigure(), which otherwise
resets the cache environment elements to NULL.

As a result, half-configured states can emerge, but the problem of not being able to build & save the cache
(which needs has_qgis()) is communicated to the user. Also the individual fails are reported. Moreover, the
half-configured state still allows some functionality, and also the hardening in this commit ensures a
graceful, non-interrupted execution of qgis_configure().