qgis_*able_plugins(): generalize message about cache rebuilding *
Sometimes you want to enable or disable a plugin but it doesn't work (because of some plugin error). In that
case cache rebuilding may often not be needed, but anyway the message 'Rebuilding cache to reflect these
changes' is awkward since no changes were made by R. However it is possible that a plugin has been removed in
QGIS, during the R session, without qgisprocess knowing about that. To allow for these situations, it's best
to indeed always rebuild the cache after qgis_*able_plugins() has called qgis_process plugins, be it
successful or not. So that is the current behaviour already; the only problem was the message. This is now
fixed.