qgis_using_json_output(): handle cache-update & json_input interaction *
- Always update the cache value if option or env var is set for json_output; this brings effect of the user
setting more on par with effect of json_input setting. This also includes an immediate effect of the option
setting on the return value of qgis_using_json_output(), just like it already was the case in
qgis_using_json_input().

- If json_output is set by the user, don't follow it if it's set as FALSE AND the user has explicitly set
json_input as TRUE. The latter always triggers json_output as TRUE in backend qgis_process.