qgis_configure(): add cache condition wrt use_json_output element *
During package configuration, reconfigure the use_json_output cache value if the corresponding option/envvar
is set by the user and resolves to the opposite of the cache value. The 'resolving' part of the user setting
includes the check for a conflicting explicit setting for json_input vs json_output, just as resolving is done in
qgis_using_json_output().