qgis_configure(): capture cache conflict use_json_output vs JSON input setting *
If the json_output user setting (ie option or env var) is empty, but the json_input user setting is not empty
AND set as TRUE in an acceptable way (i.e. minimum QGIS version requirement satified), then the
use_json_output cache value must be TRUE. If not, the cache must be reconfigured and the cache file rewritten.