qgis_using_json_output(): let query=TRUE always return TRUE in absence of user settings *
This was already the case on Windows and macOS, and this commit extends this approach to all systems. AFAIK it
was never clear in the first place why at some point it didn't work on GHA Ubuntu runner (the reason for
testing '--json' in case of non-win & non-mac), and I've never witnessed such behaviour in Linux or GHA. This
commit causes a speed gain up to several seconds for the query=TRUE case.

Note that the test mimicking a situation where --json would not work, was only relevant in the previous state
and only on Linux, so dropping these expectations.