test-compat-terra.R: relax regex for R-oldrel on macOS *
Following failure was seen in r-spatial.r-universe.dev, but not in
https://cran.r-project.org/web/checks/check_results_qgisprocess.html :

Running the tests in ‘tests/testthat.R’ failed.
Last 13 lines of output:
    'test-qgis-state.R:145:3', 'test-qgisprocess-deprecated.R:17:3'
  • Test takes ~1 hr to run (1): 'test-qgis-help.R:53:3'

  ══ Failed tests ════════════════════════════════════════════════════════════════
  ── Failure ('test-compat-terra.R:105:3'): terra argument coercers work for locally created SpatVector ──
  as_qgis_argument\(obj\[1, \], qgis_argument_spec\(qgis_type = "point"\)\) does not match "^[\\de\\+]+,[\\de\\+]+\\[\\w+:\\d+\\]$".
  Actual value: "7e\+05,7e\+05"
  Backtrace:
      ▆
   1. └─testthat::expect_match(...) at test-compat-terra.R:105:3
   2.   └─testthat:::expect_match_(...)

  [ FAIL 1 | WARN 6 | SKIP 59 | PASS 261 ]
  Error: Test failures
  Execution halted

This commit relaxes the regex match for qgis argument type 'point' in case of SpatVector, specifically in
R < 4.3 on macOS.

So this is an extension to 584a3db. It is because in the GHA log only the last 13 lines of output are displayed.

Notably, it appears that on macOS R-oldrel `terra::crs(x, describe = TRUE)[, c("authority", "code")]` does not
return the same as in other systems / R-releases; instead this result apparently has at least one NA value.