test-compat-sf.R: skip ESPG-failures *
The r-universe runners for R-oldrel on macOS seem somewhat less reliable, relating to EPSG-related failures,
to be considered as some setup side-effect.

4 failures were observed that are specific to https://github.com/r-universe/inbo.

We skip these tests if they produce a false positive in r-universe. It appears that our own GHA 'R-oldrel on
macOS' job doesn't have the problem, and it was also not observed in https://github.com/r-universe/r-spatial.